#5 refactor: hub - dto 디렉토리 세부화
