--- hub-service/src/main/java/com/fn/eureka/client/hubservice/hub/application/dto/UpdateHubRequest.java
@@ -1,4 +1,0 @@
-package com.fn.eureka.client.hubservice.hub.application.dto;
-
-public class UpdateHubRequest {
-}
