+++ hub-service/src/main/java/com/fn/eureka/client/hubservice/hub/application/dto/request/CreateHubRequest.java
@@ -1,0 +1,15 @@
+package com.fn.eureka.client.hubservice.hub.application.dto.request;
+
+import java.util.UUID;
+
+import com.fn.eureka.client.hubservice.hub.domain.HubType;
+
+import lombok.Getter;
+
+@Getter
+public class CreateHubRequest {
+	private String hubName;
+	private String hubAddress;
+	private HubType hubType;
+	private UUID hubManagerId;
+}
