+++ hub-service/src/main/java/com/fn/eureka/client/hubservice/hub/application/dto/request/UpdateHubRequest.java
@@ -1,0 +1,4 @@
+package com.fn.eureka.client.hubservice.hub.application.dto.request;
+
+public class UpdateHubRequest {
+}
