+++ hub-service/src/main/java/com/fn/eureka/client/hubservice/hub/application/dto/response/PageInfo.java
@@ -1,0 +1,13 @@
+package com.fn.eureka.client.hubservice.hub.application.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class PageInfo {
+	private long page;
+	private long size;
+	private long totalPage;
+	private long totalElement;
+}
